third_party/blink/web_tests/virtual/idna-2008/external/wpt/url/a-element-xhtml_include=javascript-expected.txt

This is a testharness.js-based test.
PASS Loading data…
PASS Parsing: <javascript:/example.com/> against <http://example.org/foo/bar>
PASS Parsing: <javascript:example.com/> against <http://example.org/foo/bar>
PASS Parsing: <javascript:/example.com/> against <about:blank>
PASS Parsing: <javascript:example.com/> against <about:blank>
FAIL Parsing: <javascript:/../> against <about:blank> assert_equals: href expected "javascript:/" but got "javascript:/../"
Harness: the test ran to completion.
